requirements.txt
alembic==0.6.6
amqp==2.2.2
attrs==17.4.0
billiard==3.5.0.3
bson==0.5.2
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.3.9
coverage==4.0a1
defusedxml==0.5.0
Flask==0.10.1
Flask-Cors==3.0.4
Flask-Login==0.2.11
Flask-Migrate==1.2.0
Flask-OpenID==1.2.5
Flask-Script==2.0.5
Flask-SQLAlchemy==1.0
Flask-Testing==0.4.2
Flask-WTF==0.10.2
Genshi==0.7
gunicorn==19.0.0
idna==2.6
itchat==1.3.10
itsdangerous==0.24
Jinja2==2.10
kombu==4.1.0
Mako==1.0.0
Markdown==2.6.11
MarkupSafe==0.23
mongoengine==0.15.0
numpy==1.14.3
opencv-python==3.4.0.14
Pillow==5.0.0
pluggy==0.6.0
py==1.5.2
Pygments==2.2.0
pymongo==3.6.1
pypng==0.0.18
PyQRCode==1.2.1
pytest==3.4.2
python-dateutil==2.7.0
python-editor==1.0.3
python-openid==2.2.5
python3-openid==3.1.0
pytz==2018.3
redis==2.10.6
requests==2.18.4
six==1.11.0
SQLAlchemy==0.9.4
urllib3==1.22
vine==1.1.4
Werkzeug==0.14.1
Whoosh==2.7.4
WTForms==2.0.1
